backend-ai/routes/sar_translation.py
from fastapi import APIRouter, File, UploadFile, HTTPException
from pydantic import BaseModel
import time
import random
import uuid
from PIL import Image
import numpy as np
import io

router = APIRouter()


class SARTranslationResponse(BaseModel):
    success: bool
    data: dict | None = None
    error: dict | None = None


@router.post("/sar-translate")
async def sar_to_optical_translation(image: UploadFile = File(...)):
    """
    Translate SAR imagery to optical-like image using CycleGAN.
    
    In production, this would use MindSpore CycleGAN model trained on
    Sentinel-1 SAR and Sentinel-2 optical pairs.
    """
    try:
        start_time = time.time()
        
        # Validate image
        if not image.content_type.startswith("image/"):
            raise HTTPException(status_code=400, detail="File must be an image")
        
        # Read and process image (mock)
        contents = await image.read()
        img = Image.open(io.BytesIO(contents))
        
        # Mock translation (in production: run CycleGAN inference)
        # For now, just save the image as "translated"
        filename = f"{uuid.uuid4()}.png"
        output_path = f"static/translated/{filename}"
        img.save(output_path)
        
        processing_time = time.time() - start_time
        
        # Mock SSIM score (Structural Similarity Index)
        ssim = round(0.85 + random.uniform(0, 0.1), 3)
        
        return SARTranslationResponse(
            success=True,
            data={
                "imageUrl": f"/static/translated/{filename}",
                "ssim": ssim,
                "processingTime": round(processing_time, 2),
                "model": "CycleGAN (MindSpore)",
                "resolution": f"{img.width}x{img.height}",
            }
        )
        
    except Exception as e:
        return SARTranslationResponse(
            success=False,
            error={
                "code": "AI_001",
                "message": "SAR translation failed",
                "details": str(e),
            }
        )
