backend-ai/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.1.0
httpx>=0.25.0
